Make connectTimeout configurable in mySQL conn URL.
Without this setting, if someone POSTs a connector config with
an unreachable db hostname, the verify routine of the mySQL
connector will hang indefinitely.  Interestingly, this
prevents other connectors from being submitted (a synchronization
issue that this does not address).

With this patch, though, the value for the `connect.timeout.ms`
connector property will be set as the `connectTimeout` URL parameter,
which, if nonzero, will cause the connection to eventually
time out.

We're currently using v0.7.2 but am submitting this patch against
master.  If accepted, I'd like to backpatch it to the v0.7x timeline
as well.